requirements.txt
Django==1.7
Markdown==2.5
Werkzeug==0.9.6
django-filter==0.7
django-rest-framework-mongoengine==1.5.1
djangorestframework==2.4.2
mongoengine==0.8.7
pymongo==2.7.2
wsgiref==0.1.2
